moved struct to another file
