+++ tests/Greenspun/structs/EvalContext.h
@@ -1,0 +1,24 @@
+#include <iostream>
+
+#include "Pregel/Algos/VertexAccumulators/Greenspun/Primitives.h"
+
+struct MyEvalContext : PrimEvalContext {
+  std::string const& getThisId() const override { std::abort(); }
+
+  void getAccumulatorValue(std::string_view id, VPackBuilder& result) const override {
+    std::abort();
+  }
+
+  void updateAccumulator(std::string_view accumId, std::string_view vertexId,
+                         VPackSlice value) override {
+    std::abort();
+  }
+
+  void setAccumulator(std::string_view accumId, VPackSlice value) override {
+    std::abort();
+  }
+
+  EvalResult enumerateEdges(std::function<EvalResult(VPackSlice edge)> cb) const override {
+    std::abort();
+  }
+};
